requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
jinja2==3.1.2
openai==1.3.5
weasyprint==60.1
python-dotenv==1.0.0
pydantic==2.4.2
httpx==0.24.1 